physical_ue_proxy/requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0

Flask~=2.0.0
six~=1.14.0
PyYAML~=5.3.1
aiohttp~=3.7.4.post0